feat(components-pipeline-1): move button to atoms
